+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 4.0-M2
 
 * #363 Upgrade to JUnit 5.13.4
+* #364 "--config=stdin:xxx" - STDIN as an extra config source
 
 ## 4.0-M1